agent_project/requirement.txt
markdown-it-py
openai
boto3
python-dotenv
tqdm